refactor : separate components from screen, add web view on desktop
