+++ src/app/index.css
@@ -50,6 +50,15 @@
 }
 
 @layer utilities {
+  .container-mobile {
+    max-width: 500px;
+    margin: 0 auto;
+  }
+
+  .container-width {
+    max-width: 500px;
+  }
+
   .scrollbar-hide::-webkit-scrollbar {
     display: none;
   }
@@ -61,5 +70,16 @@
 
   .box-shadow-buttonLg {
     box-shadow: 0px 4px 10px 0px rgba(0, 0, 0, 0.2);
+  }
+
+  .box-shadow-dock {
+    box-shadow: 0px -2px 20px 0px rgba(0, 0, 0, 0.1);
+  }
+}
+
+@media (prefers-color-scheme: dark) {
+  :root {
+    color-scheme: dark;
+    --color-base-100: #fff;
   }
 }
