+++ src/widgets/home/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as HomeContainer } from './HomeContainer'
+export { default as GroupCarousel } from './GroupCarousel'
